Prevent multiple DX warnings due to resource barrier.
This silences a never ending stream of "DX WARNING: ID3D12CommandList::ResourceBarrier: Called on the same subresource(s) of Resource(XXX) in separate Barrier Descs which is inefficient and likely unintentional".
There is already a todo about needing multiple resource flags which may be the proper long term fix. This however fixes it in the mean time.