Allow providing separate metal/rough textures.
This is required to get best results from using compressed textures as it allows for an individual texture to be provided for each channel instead of a combined one. There is expected work within the gltf specification to support this kind of thing as part of the spec in the future however until that is included then this can be used as a work around.